WIP: Simple usage of contracts in TR
Stores the user-specified contract in a syntax property on the provide'd
identifier. Later, when contracts are potentially generated for provided
terms, the user's contract is added to the provided value.

For short-term simplicity, having a user contract disables any attempt
to cooperate with optimizations performed by the contract library.

Some remaining problems:
- have to write ->/c instead of -> in contracts
- what if the contract refers to a user-level identifier?
- error messages mention and/c which the user didn't write and the type
  annotation doesn't correspond to
- user contracts aren't enforced in typed-typed interaction