WIP: Add contract type Con and typecheck contract expressions
Arrow contracts expand into a big mess, so we take a similar approach
to typechecking classes: mark the surface syntax, let it expand, find
the expanded+marked bits and then typecheck. Note that traversing the
expansion doesn't necessarily visit the domains in the same order as
the user wrote them, so we have to maintain the order ourselves.