WIP: Make combinators that consume/produce flat contracts use FlatCon
Additionally:
- Fixed tests to refer to FlatCon
- Added a match expander for matching on Con or FlatCon, since they're
  treated the same in check-unit
- Special-cased the inference for FlatCon and Con

The last point is the most surprising. The subtype case for inference
doesn't seem to work when the "above" type is polymorphic. I was hoping
to be able to use a (FlatCon Any) where a (Con a) was asked for (and
the FlatCon and Con are related covariantly) but the inference didn't
work as desired.