WIP, needs a lot of cleanup
Lots of redundancies in the syntax properties for the contracts. Also
need to look into cleaning up error messages in ->i since they'll report
based on the modified (begin ...) we insert for each contract.

Properties for ->i need to be on each of the contracts, not on the id
or on the id+ctc pairs. Also need to get the expanded version of the
dependencies, since ->i futzes with those (and doesn't use the same
identifier that the user wrote).

Ignore property on ->i is tricky, since we still want it to be
typechecked as the rest of expressions are. Not sure how to get around
the ignore, well, being ignored at the top-level of a module and still
have the contract be typechecked.