WIP: Internalize predicate/flat contract relationship (subtype + infer)
TODO: remove special casing, if possible, from check-contract. Before,
this relationship was handled inside check-contract. It still is as of
this commit, but that should be gone after I've confirmed this handles
those cases and then rebasing.

Unclear if the PredicateFilter <: FlatCon case is correct. It seems
fine at first, e.g. something that checks an exact-integer? should be
usable in place of something that requires real?. But functions are
contravariant in the domain, so I hope I'm not missing something.