Keep non-Any->Any predicates out of contracts
Because we don't yet refine the type of a monitored term, we can't allow
unrestricted predicates as contracts. If we did, that would allow the
contract to test values that don't agree with expected input type.

Consider:

k = (->/c (lambda ([x : Real]) (> x 10)) (lambda (x) #t))
e = (lambda (x) x)

Once e is monitored by k, applying the resulting expression to a string
would allow that string to get to the body of the lambda.