WIP: Convert everything to the new two-param Con
Things to still figure out:
- Why does the constraint generation not look like the Param case?
- or/c for higher-order contracts
- Support for expected types in contract typing rules

Because poly-dots only allows for a single dotted variable, list/c needs
its own typing rule and can't be expressed as a simple function type.

Without intersections, we can't monitor an Integer function with
>/c. Some of the tests needed to be changed then.

We also can't not/c a (Con Any String) in the domain of a function
contract -- this produces a (Con Any Any) which ends up making the input
and the output types require Any in their domain. With the current type
for string-len/c, we have to guard it with a type test to get the right
type.