Combine monitored id's ty w/contract's output ty
A contract (->/c exact-integer? exact-integer?) has an output type
of (-> Any Integer). We don't want that as the monitored id's type,
so we need a way to safely combine the monitored id's type and the
output term's type (we can't just use the monitored id's type).

A pairwise intersect appears to be OK. It's also a lower bound of
the two types wrt the "information order" (aka naive subtype
order).

Note: This doesn't work well with contracts on polymorphic types,
like the ones for sorted in the Acquire benchmark. That's on the
TODO list.

Also, this commit fixes non-TypeProp function to contract coercion.