Add plain, unannotated provide/contract etc.
Typechecking provide/contract no longer requires assistance from
Racket.

Because of odd behavior in contract-out's expansion, we provide
a contract-out that delegates to provide/contract. This means error
messages will mention provide/contract even though the user wrote
contract-out.

The type added for coerce-contract is concerning because it will have to
encode a lot of the rules that are already used for typechecking
contracts, but it's not clear how to do better (is there a way to make
it a type rule?). Importantly, we can't make things coercible to a
contract a subtype of flat contract type; if we did, we would be able to
apply non-functions (e.g. strings, reals, regexps, etc.) as if they were
functions.